backend/requirements.txt
sentence-transformers==2.2.2
langchain
flask
python-dotenv
langchain-pinecone
langchain-community
langchain-google-genai
langchain-huggingface
pdfminer.six
pinecone-client
tqdm
